+++ src/mplchart/primitives/__init__.py
@@ -6,8 +6,10 @@ from .volume import Volume
 from .ohlc import OHLC
 from .candlesticks import Candlesticks
 from .peaks import Peaks
+from .stripes import Stripes
 from .lineplot import LinePlot
 from .areaplot import AreaPlot
 from .barplot import BarPlot
+